Changed folder name for ui folder
