--- gui/ui.py
@@ -1,32 +1,0 @@
-#!/usr/bin/env python
-import sys
-from PyQt5.QtWidgets import QApplication, QWidget, QMainWindow, QLabel
-from PyQt5.QtGui import QIcon
-
-class Gui(QMainWindow):
-    '''Create a window and display the title of the project in the center'''
-
-    def __init__(self):
-        super().__init__()
-        self.title = ''
-        self.left = 10
-        self.top = 10
-        self.width = 440
-        self.height = 280
-        self.initUI()
-
-    def initUI(self):
-        self.setWindowTitle(self.title)
-        self.setGeometry(self.left, self.top, self.width, self.height)
-
-        label = QLabel('SIG Blockchain Project', self)
-        label.move(150,0)
-        label.adjustSize()
-
-
-        self.show()
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    ex = Gui()
-    sys.exit(app.exec_())
